modified script
- script name convention changed
